requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
python-dotenv==1.0.1
tweepy==4.14.0
transformers==4.44.2
pydantic==2.9.2
pytest==7.4.3
httpx==0.25.2